+++ deployment/tokens/configure_tokens_for_transfers.go
@@ -854,3 +854,60 @@ func LegacyRateLimitsForAutoMigrate[R any, CCV any](
 		Inbound:  inboundLegacy,
 	}, nil
 }
+
+func ResolveTokenFeeAdapter(e cldf.Environment, sel uint64, poolRef datastore.AddressRef) (TokenFeeAdapter, error) {
+	registry := GetTokenAdapterRegistry()
+
+	fam, err := chain_selectors.GetSelectorFamily(sel)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get chain selector family for selector %d: %w", sel, err)
+	}
+
+	ref, err := ResolveTokenPoolRef(e, registry, sel, poolRef)
+	if err != nil {
+		return nil, fmt.Errorf("failed to resolve token pool ref: %w", err)
+	}
+
+	tokAdapter, ok := registry.GetTokenAdapter(fam, ref.Version)
+	if !ok {
+		return nil, fmt.Errorf("no token adapter found for chain family %s and pool ref %s", datastore_utils.SprintRef(poolRef), fam)
+	}
+
+	feeAdapter, ok := tokAdapter.(TokenFeeAdapter)
+	if !ok {
+		return nil, fmt.Errorf("token adapter for chain family %s and pool ref %s does not implement TokenFeeAdapter", datastore_utils.SprintRef(poolRef), fam)
+	}
+
+	return feeAdapter, nil
+}
+
+// GetDefaultChainAgnosticTokenTransferFeeConfig returns sensible default token transfer fee configuration
+// for the given source and destination chain selectors. It may be overridden via the optional overrides
+// argument, allowing the caller to customize specific fields of the returned config.
+func GetDefaultChainAgnosticTokenTransferFeeConfig(src uint64, dst uint64, overrides ...func(*TokenTransferFeeConfig)) TokenTransferFeeConfig {
+	minFeeUSDCents := uint32(25)
+
+	// NOTE: we validate that src != dst so only one of these if statements will execute
+	if src == chain_selectors.ETHEREUM_MAINNET.Selector {
+		minFeeUSDCents = 50
+	}
+	if dst == chain_selectors.ETHEREUM_MAINNET.Selector {
+		minFeeUSDCents = 150
+	}
+
+	cfg := TokenTransferFeeConfig{
+		DefaultFinalityTransferFeeBps: 0,
+		CustomFinalityTransferFeeBps:  0,
+		DefaultFinalityFeeUSDCents:    minFeeUSDCents,
+		CustomFinalityFeeUSDCents:     minFeeUSDCents,
+		DestBytesOverhead:             32,
+		DestGasOverhead:               90_000,
+		IsEnabled:                     true,
+	}
+
+	for _, override := range overrides {
+		override(&cfg)
+	}
+
+	return cfg
+}
