+++ integration-tests/deployment/set_token_transfer_fee_test.go
@@ -9,30 +9,25 @@ import (
 	"github.com/smartcontractkit/chainlink-deployments-framework/operations"
 	"github.com/stretchr/testify/require"
 
-	datastore_utils_evm "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/utils/datastore"
 	evmadaptersV1_0_0 "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v1_0_0/adapters"
-	bnmERC20ops "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v1_0_0/operations/burn_mint_erc20"
 	evmadaptersV1_6_0 "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v1_6_0/adapters"
 	onrampV1_6 "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v1_6_0/operations/onramp"
 	evmseqV1_6_0 "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v1_6_0/sequences"
 	evmadaptersV2_0_0 "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v2_0_0/adapters"
 	fqV2_0 "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v2_0_0/operations/fee_quoter"
-	tpopsV2_0_0 "github.com/smartcontractkit/chainlink-ccip/chains/evm/deployment/v2_0_0/operations/token_pool"
 	_ "github.com/smartcontractkit/chainlink-ccip/chains/solana/deployment/v1_0_0/adapters"
+
 	soladaptersV1_6_0 "github.com/smartcontractkit/chainlink-ccip/chains/solana/deployment/v1_6_0/adapters"
 	"github.com/smartcontractkit/chainlink-ccip/chains/solana/deployment/v1_6_0/operations/router"
 	solseqV1_6_0 "github.com/smartcontractkit/chainlink-ccip/chains/solana/deployment/v1_6_0/sequences"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/deploy"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/fees"
-	"github.com/smartcontractkit/chainlink-ccip/deployment/finality"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/lanes"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/testhelpers"
-	"github.com/smartcontractkit/chainlink-ccip/deployment/tokens"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/utils"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/utils/changesets"
 	datastore_utils "github.com/smartcontractkit/chainlink-ccip/deployment/utils/datastore"
 	"github.com/smartcontractkit/chainlink-ccip/deployment/utils/mcms"
-	"github.com/smartcontractkit/chainlink-deployments-framework/chain/evm/operations/contract"
 )
 
 func TestSetTokenTransferFeeV1_6_0(t *testing.T) {
@@ -497,323 +492,4 @@ func TestSetTokenTransferFeeV2_0_0(t *testing.T) {
 	dstCfg, err = evmFeesAdapterV2_0_0.GetOnchainTokenTransferFeeConfig(e.OperationsBundle, e.BlockChains, dstFQ, dst, src, dstTokenAddress)
 	require.NoError(t, err)
 	require.False(t, dstCfg.IsEnabled)
-}
-
-func TestSetTokenPoolTokenTransferFeeV2_0_0(t *testing.T) {
-	// Define EVM chains
-	evmChainSelA := chainsel.TEST_90000001.Selector
-	evmChainSelB := chainsel.TEST_90000002.Selector
-	evmChainSelC := chainsel.TEST_90000003.Selector
-
-	// Setup test environment
-	env, err := environment.New(t.Context(), environment.WithEVMSimulated(t, []uint64{evmChainSelA, evmChainSelB, evmChainSelC}))
-	require.NoError(t, err)
-
-	// Get chains
-	evmChainA, ok := env.BlockChains.EVMChains()[evmChainSelA]
-	require.True(t, ok)
-	evmChainB, ok := env.BlockChains.EVMChains()[evmChainSelB]
-	require.True(t, ok)
-	evmChainC, ok := env.BlockChains.EVMChains()[evmChainSelC]
-	require.True(t, ok)
-
-	// Configure deployment registry
-	deployRegistry := deploy.GetRegistry()
-	deployRegistry.RegisterDeployer(chainsel.FamilyEVM, deploy.MCMSVersion, &evmadaptersV1_0_0.EVMDeployer{})
-	chainsToDeploy := map[uint64]deploy.ContractDeploymentConfigPerChain{
-		evmChainSelA: NewDefaultDeploymentConfigForEVM(utils.Version_1_6_0),
-		evmChainSelB: NewDefaultDeploymentConfigForEVM(utils.Version_1_6_0),
-		evmChainSelC: NewDefaultDeploymentConfigForEVM(utils.Version_1_6_0),
-	}
-
-	// Deploy contracts
-	output, err := deploy.DeployContracts(deployRegistry).Apply(*env, deploy.ContractDeploymentConfig{Chains: chainsToDeploy, MCMS: mcms.Input{}})
-	require.NoError(t, err)
-	MergeAddresses(t, env, output.DataStore)
-
-	// Deploy MCMS and transfer ownership
-	DeployMCMS(t, env, evmChainSelA, []string{utils.CLLQualifier})
-	DeployMCMS(t, env, evmChainSelB, []string{utils.CLLQualifier})
-	DeployMCMS(t, env, evmChainSelC, []string{utils.CLLQualifier})
-	EVMTransferOwnership(t, env, evmChainSelA)
-	EVMTransferOwnership(t, env, evmChainSelB)
-	EVMTransferOwnership(t, env, evmChainSelC)
-
-	// Define a full pool mesh between all EVM chains
-	disabledOutboundRL := tokens.RateLimiterConfigFloatInput{IsEnabled: false}
-	tokenExpansionInput := map[uint64]tokens.TokenExpansionInputPerChain{
-		evmChainSelA: {
-			TokenPoolVersion: utils.Version_2_0_0,
-			DeployTokenInput: &tokens.DeployTokenInput{
-				Type:     bnmERC20ops.ContractType,
-				Decimals: 18,
-				Symbol:   "TEST_TOKEN_B",
-				Name:     "Test Token B",
-				PreMint:  nil, // no pre-mint
-				Supply:   nil, // unlimited
-			},
-			DeployTokenPoolInput: &tokens.DeployTokenPoolInput{
-				PoolType: utils.BurnMintTokenPool.String(),
-			},
-			TokenTransferConfig: &tokens.TokenTransferConfig{
-				RemoteChains: map[uint64]tokens.RemoteChainConfig[*datastore.AddressRef, datastore.AddressRef]{
-					evmChainSelB: {OutboundRateLimiterConfig: &disabledOutboundRL},
-					evmChainSelC: {OutboundRateLimiterConfig: &disabledOutboundRL},
-				},
-			},
-		},
-		evmChainSelB: {
-			TokenPoolVersion: utils.Version_2_0_0,
-			DeployTokenInput: &tokens.DeployTokenInput{
-				Type:     bnmERC20ops.ContractType,
-				Decimals: 18,
-				Symbol:   "TEST_TOKEN_B",
-				Name:     "Test Token B",
-				PreMint:  nil, // no pre-mint
-				Supply:   nil, // unlimited
-			},
-			DeployTokenPoolInput: &tokens.DeployTokenPoolInput{
-				PoolType: utils.BurnMintTokenPool.String(),
-			},
-			TokenTransferConfig: &tokens.TokenTransferConfig{
-				RemoteChains: map[uint64]tokens.RemoteChainConfig[*datastore.AddressRef, datastore.AddressRef]{
-					evmChainSelA: {OutboundRateLimiterConfig: &disabledOutboundRL},
-					evmChainSelC: {OutboundRateLimiterConfig: &disabledOutboundRL},
-				},
-			},
-		},
-		evmChainSelC: {
-			TokenPoolVersion: utils.Version_2_0_0,
-			DeployTokenInput: &tokens.DeployTokenInput{
-				Type:     bnmERC20ops.ContractType,
-				Decimals: 18,
-				Symbol:   "TEST_TOKEN_C",
-				Name:     "Test Token C",
-				PreMint:  nil, // no pre-mint
-				Supply:   nil, // unlimited
-			},
-			DeployTokenPoolInput: &tokens.DeployTokenPoolInput{
-				PoolType: utils.BurnMintTokenPool.String(),
-			},
-			TokenTransferConfig: &tokens.TokenTransferConfig{
-				RemoteChains: map[uint64]tokens.RemoteChainConfig[*datastore.AddressRef, datastore.AddressRef]{
-					evmChainSelA: {OutboundRateLimiterConfig: &disabledOutboundRL},
-					evmChainSelB: {OutboundRateLimiterConfig: &disabledOutboundRL},
-				},
-			},
-		},
-	}
-
-	// Deploy the pool mesh
-	output, err = tokens.TokenExpansion().Apply(*env, tokens.TokenExpansionInput{
-		TokenExpansionInputPerChain: tokenExpansionInput,
-		ChainAdapterVersion:         utils.Version_2_0_0,
-		MCMS:                        NewDefaultInputForMCMS("Token Expansion"),
-	})
-	require.NoError(t, err)
-	MergeAddresses(t, env, output.DataStore)
-
-	// Fetch the pool addresses from the datastore
-	fltrA := datastore.AddressRef{Type: datastore.ContractType(tokenExpansionInput[evmChainSelA].DeployTokenPoolInput.PoolType), Version: tokenExpansionInput[evmChainSelA].DeployTokenPoolInput.TokenPoolVersion}
-	poolA, err := datastore_utils.FindAndFormatRef(env.DataStore, fltrA, evmChainSelA, datastore_utils_evm.ToEVMAddress)
-	require.NoError(t, err)
-	fltrB := datastore.AddressRef{Type: datastore.ContractType(tokenExpansionInput[evmChainSelB].DeployTokenPoolInput.PoolType), Version: tokenExpansionInput[evmChainSelB].DeployTokenPoolInput.TokenPoolVersion}
-	poolB, err := datastore_utils.FindAndFormatRef(env.DataStore, fltrB, evmChainSelB, datastore_utils_evm.ToEVMAddress)
-	require.NoError(t, err)
-	fltrC := datastore.AddressRef{Type: datastore.ContractType(tokenExpansionInput[evmChainSelC].DeployTokenPoolInput.PoolType), Version: tokenExpansionInput[evmChainSelC].DeployTokenPoolInput.TokenPoolVersion}
-	poolC, err := datastore_utils.FindAndFormatRef(env.DataStore, fltrC, evmChainSelC, datastore_utils_evm.ToEVMAddress)
-	require.NoError(t, err)
-
-	// Set the token pool token transfer fee config on all pools
-	output, err = tokens.
-		SetTokenTransferFee().
-		Apply(*env, tokens.SetTokenTransferFeeInput{
-			Version: utils.Version_2_0_0,
-			MCMS:    NewDefaultInputForMCMS("Set token transfer fee"),
-			Args: []tokens.TokenTransferFeeForSrc{
-				{
-					Selector: evmChainSelA,
-					TokenPools: []tokens.TokenTransferFeeForPool{
-						{
-							PoolAddress:           poolA.Hex(),
-							AllowedFinalityConfig: finality.Config{BlockDepth: 12},
-							Destinations: []tokens.TokenTransferFeeForDst{
-								{
-									IsReset:  false,
-									Selector: evmChainSelB,
-									Settings: tokens.UnresolvedTokenTransferFeeArgs{
-										DefaultFinalityTransferFeeBps: utils.NewOptional(uint16(100)),
-									},
-								},
-								{
-									IsReset:  false,
-									Selector: evmChainSelC,
-									Settings: tokens.UnresolvedTokenTransferFeeArgs{
-										CustomFinalityTransferFeeBps: utils.NewOptional(uint16(10)),
-									},
-								},
-							},
-						},
-					},
-				},
-				{
-					Selector: evmChainSelB,
-					TokenPools: []tokens.TokenTransferFeeForPool{
-						{
-							PoolAddress:           poolB.Hex(),
-							AllowedFinalityConfig: finality.Config{},
-							Destinations: []tokens.TokenTransferFeeForDst{
-								{
-									IsReset:  false,
-									Selector: evmChainSelA,
-									Settings: tokens.UnresolvedTokenTransferFeeArgs{
-										DefaultFinalityFeeUSDCents: utils.NewOptional(uint32(200)),
-									},
-								},
-								{
-									IsReset:  false,
-									Selector: evmChainSelC,
-									Settings: tokens.UnresolvedTokenTransferFeeArgs{
-										CustomFinalityFeeUSDCents: utils.NewOptional(uint32(20)),
-									},
-								},
-							},
-						},
-					},
-				},
-				{
-					Selector: evmChainSelC,
-					TokenPools: []tokens.TokenTransferFeeForPool{
-						{
-							PoolAddress:           poolC.Hex(),
-							AllowedFinalityConfig: finality.Config{WaitForFinality: true},
-							Destinations: []tokens.TokenTransferFeeForDst{
-								{
-									IsReset:  false,
-									Selector: evmChainSelA,
-									Settings: tokens.UnresolvedTokenTransferFeeArgs{
-										DestBytesOverhead: utils.NewOptional(uint32(100_000)),
-										DestGasOverhead:   utils.NewOptional(uint32(10_000)),
-									},
-								},
-								{
-									IsReset:  false,
-									Selector: evmChainSelB,
-									Settings: tokens.UnresolvedTokenTransferFeeArgs{
-										DestBytesOverhead: utils.NewOptional(uint32(200_000)),
-										DestGasOverhead:   utils.NewOptional(uint32(20_000)),
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-		})
-	require.NoError(t, err)
-	testhelpers.ProcessTimelockProposals(t, *env, output.MCMSTimelockProposals, false)
-
-	// Get the v2.0 token adapter
-	tokensV2 := evmadaptersV2_0_0.TokenAdapter{}
-
-	// Reset the operation cache
-	env.OperationsBundle = operations.NewBundle(t.Context, env.OperationsBundle.Logger, operations.NewMemoryReporter())
-
-	// Check allowed finality config for pool A (BlockDepth=12 was set)
-	reportA, err := operations.ExecuteOperation(
-		env.OperationsBundle, tpopsV2_0_0.GetAllowedFinalityConfig, evmChainA,
-		contract.FunctionInput[struct{}]{ChainSelector: evmChainSelA, Address: poolA, Args: struct{}{}},
-	)
-	require.NoError(t, err)
-	require.Equal(t, finality.Config{BlockDepth: 12}.Raw(), reportA.Output)
-
-	// Check allowed finality config for pool B (not set, defaults to WaitForFinality)
-	reportB, err := operations.ExecuteOperation(
-		env.OperationsBundle, tpopsV2_0_0.GetAllowedFinalityConfig, evmChainB,
-		contract.FunctionInput[struct{}]{ChainSelector: evmChainSelB, Address: poolB, Args: struct{}{}},
-	)
-	require.NoError(t, err)
-	require.Equal(t, finality.RawWaitForFinality, reportB.Output)
-
-	// Check allowed finality config for pool C (explicitly set to 0, i.e. WaitForFinality)
-	reportC, err := operations.ExecuteOperation(
-		env.OperationsBundle, tpopsV2_0_0.GetAllowedFinalityConfig, evmChainC,
-		contract.FunctionInput[struct{}]{ChainSelector: evmChainSelC, Address: poolC, Args: struct{}{}},
-	)
-	require.NoError(t, err)
-	require.Equal(t, finality.RawWaitForFinality, reportC.Output)
-
-	// Query the on-chain config for both pools
-	cfgAB, err := tokensV2.GetOnchainTokenTransferFeeConfig(*env, poolA.Hex(), evmChainSelA, evmChainSelB)
-	require.NoError(t, err)
-	cfgAC, err := tokensV2.GetOnchainTokenTransferFeeConfig(*env, poolA.Hex(), evmChainSelA, evmChainSelC)
-	require.NoError(t, err)
-	cfgBA, err := tokensV2.GetOnchainTokenTransferFeeConfig(*env, poolB.Hex(), evmChainSelB, evmChainSelA)
-	require.NoError(t, err)
-	cfgBC, err := tokensV2.GetOnchainTokenTransferFeeConfig(*env, poolB.Hex(), evmChainSelB, evmChainSelC)
-	require.NoError(t, err)
-	cfgCA, err := tokensV2.GetOnchainTokenTransferFeeConfig(*env, poolC.Hex(), evmChainSelC, evmChainSelA)
-	require.NoError(t, err)
-	cfgCB, err := tokensV2.GetOnchainTokenTransferFeeConfig(*env, poolC.Hex(), evmChainSelC, evmChainSelB)
-	require.NoError(t, err)
-
-	// Confirm that the configs for A->B match what was set, and that sensible defaults are applied for fields that were not set
-	defaultsAB := tokens.GetDefaultChainAgnosticTokenTransferFeeConfig(evmChainSelA, evmChainSelB)
-	require.Equal(t, uint16(100), cfgAB.DefaultFinalityTransferFeeBps)
-	require.Equal(t, defaultsAB.CustomFinalityTransferFeeBps, cfgAB.CustomFinalityTransferFeeBps)
-	require.Equal(t, defaultsAB.DefaultFinalityFeeUSDCents, cfgAB.DefaultFinalityFeeUSDCents)
-	require.Equal(t, defaultsAB.CustomFinalityFeeUSDCents, cfgAB.CustomFinalityFeeUSDCents)
-	require.Equal(t, defaultsAB.DestBytesOverhead, cfgAB.DestBytesOverhead)
-	require.Equal(t, defaultsAB.DestGasOverhead, cfgAB.DestGasOverhead)
-	require.True(t, cfgAB.IsEnabled)
-
-	// Confirm that the configs for A->C match what was set, and that sensible defaults are applied for fields that were not set
-	defaultsAC := tokens.GetDefaultChainAgnosticTokenTransferFeeConfig(evmChainSelA, evmChainSelC)
-	require.Equal(t, defaultsAC.DefaultFinalityTransferFeeBps, cfgAC.DefaultFinalityTransferFeeBps)
-	require.Equal(t, uint16(10), cfgAC.CustomFinalityTransferFeeBps)
-	require.Equal(t, defaultsAC.DefaultFinalityFeeUSDCents, cfgAC.DefaultFinalityFeeUSDCents)
-	require.Equal(t, defaultsAC.CustomFinalityFeeUSDCents, cfgAC.CustomFinalityFeeUSDCents)
-	require.Equal(t, defaultsAC.DestBytesOverhead, cfgAC.DestBytesOverhead)
-	require.Equal(t, defaultsAC.DestGasOverhead, cfgAC.DestGasOverhead)
-	require.True(t, cfgAC.IsEnabled)
-
-	// Confirm that the configs for B->A match what was set, and that sensible defaults are applied for fields that were not set
-	defaultsBA := tokens.GetDefaultChainAgnosticTokenTransferFeeConfig(evmChainSelB, evmChainSelA)
-	require.Equal(t, defaultsBA.DefaultFinalityTransferFeeBps, cfgBA.DefaultFinalityTransferFeeBps)
-	require.Equal(t, defaultsBA.CustomFinalityTransferFeeBps, cfgBA.CustomFinalityTransferFeeBps)
-	require.Equal(t, uint32(200), cfgBA.DefaultFinalityFeeUSDCents)
-	require.Equal(t, defaultsBA.CustomFinalityFeeUSDCents, cfgBA.CustomFinalityFeeUSDCents)
-	require.Equal(t, defaultsBA.DestBytesOverhead, cfgBA.DestBytesOverhead)
-	require.Equal(t, defaultsBA.DestGasOverhead, cfgBA.DestGasOverhead)
-	require.True(t, cfgBA.IsEnabled)
-
-	// Confirm that the configs for B->C match what was set, and that sensible defaults are applied for fields that were not set
-	defaultsBC := tokens.GetDefaultChainAgnosticTokenTransferFeeConfig(evmChainSelB, evmChainSelC)
-	require.Equal(t, defaultsBC.DefaultFinalityTransferFeeBps, cfgBC.DefaultFinalityTransferFeeBps)
-	require.Equal(t, defaultsBC.CustomFinalityTransferFeeBps, cfgBC.CustomFinalityTransferFeeBps)
-	require.Equal(t, defaultsBC.DefaultFinalityFeeUSDCents, cfgBC.DefaultFinalityFeeUSDCents)
-	require.Equal(t, uint32(20), cfgBC.CustomFinalityFeeUSDCents)
-	require.Equal(t, defaultsBC.DestBytesOverhead, cfgBC.DestBytesOverhead)
-	require.Equal(t, defaultsBC.DestGasOverhead, cfgBC.DestGasOverhead)
-	require.True(t, cfgBC.IsEnabled)
-
-	// Confirm that the configs for C->A match what was set, and that sensible defaults are applied for fields that were not set
-	defaultsCA := tokens.GetDefaultChainAgnosticTokenTransferFeeConfig(evmChainSelC, evmChainSelA)
-	require.Equal(t, defaultsCA.DefaultFinalityTransferFeeBps, cfgCA.DefaultFinalityTransferFeeBps)
-	require.Equal(t, defaultsCA.CustomFinalityTransferFeeBps, cfgCA.CustomFinalityTransferFeeBps)
-	require.Equal(t, defaultsCA.DefaultFinalityFeeUSDCents, cfgCA.DefaultFinalityFeeUSDCents)
-	require.Equal(t, defaultsCA.CustomFinalityFeeUSDCents, cfgCA.CustomFinalityFeeUSDCents)
-	require.Equal(t, uint32(100_000), cfgCA.DestBytesOverhead)
-	require.Equal(t, uint32(10_000), cfgCA.DestGasOverhead)
-	require.True(t, cfgCA.IsEnabled)
-
-	// Confirm that the configs for C->B match what was set, and that sensible defaults are applied for fields that were not set
-	defaultsCB := tokens.GetDefaultChainAgnosticTokenTransferFeeConfig(evmChainSelC, evmChainSelB)
-	require.Equal(t, defaultsCB.DefaultFinalityTransferFeeBps, cfgCB.DefaultFinalityTransferFeeBps)
-	require.Equal(t, defaultsCB.CustomFinalityTransferFeeBps, cfgCB.CustomFinalityTransferFeeBps)
-	require.Equal(t, defaultsCB.DefaultFinalityFeeUSDCents, cfgCB.DefaultFinalityFeeUSDCents)
-	require.Equal(t, defaultsCB.CustomFinalityFeeUSDCents, cfgCB.CustomFinalityFeeUSDCents)
-	require.Equal(t, uint32(200_000), cfgCB.DestBytesOverhead)
-	require.Equal(t, uint32(20_000), cfgCB.DestGasOverhead)
-	require.True(t, cfgCB.IsEnabled)
 }
